@@ -13,7 +13,7 @@
     "minimum-stability": "stable",
     "prefer-stable": true,
     "require": {
-        "php": ">=8.0"
+        "php": ">=8.1"
     },
     "config": {
         "optimize-autoloader": true,
@@ -29,6 +29,6 @@
         "test": "./vendor/bin/phpunit"
     },
     "require-dev": {
-        "phpunit/phpunit": "^10.2"
+        "phpunit/phpunit": "^10.3"
     }
 }
